MAT-1355 - move dividendAccounts to topup module
