+++ staking/client/rest/query.go
@@ -3,7 +3,6 @@ package rest
 import (
 	"encoding/json"
 	"fmt"
-	"math/big"
 	"net/http"
 
 	"github.com/cosmos/cosmos-sdk/client/context"
@@ -42,28 +41,8 @@ func registerQueryRoutes(cliCtx context.CLIContext, r *mux.Router) {
 		currentProposerHandlerFn(cliCtx),
 	).Methods("GET")
 	r.HandleFunc(
-		"/staking/slash-validator",
-		slashValidatorHandlerFn(cliCtx),
-	).Methods("GET")
-	r.HandleFunc(
-		"/staking/dividend-account/{id}",
-		dividendAccountByIDHandlerFn(cliCtx),
-	).Methods("GET")
-	r.HandleFunc(
-		"/staking/dividend-account-root",
-		dividendAccountRootHandlerFn(cliCtx),
-	).Methods("GET")
-	r.HandleFunc(
 		"/staking/proposer-bonus-percent",
 		proposerBonusPercentHandlerFn(cliCtx),
-	).Methods("GET")
-	r.HandleFunc(
-		"/staking/account-proof/{id}",
-		dividendAccountProofHandlerFn(cliCtx),
-	).Methods("GET")
-	r.HandleFunc(
-		"/staking/account-proof/verify/",
-		VerifyAccountProofHandlerFn(cliCtx),
 	).Methods("GET")
 	r.HandleFunc(
 		"/staking/isoldtx",
@@ -289,134 +268,6 @@ func currentProposerHandlerFn(cliCtx context.CLIContext) http.HandlerFunc {
 	}
 }
 
-func slashValidatorHandlerFn(
-	cliCtx context.CLIContext,
-) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-
-		cliCtx, ok := rest.ParseQueryHeightOrReturnBadRequest(w, cliCtx, r)
-		if !ok {
-			return
-		}
-
-		params := r.URL.Query()
-		valID, ok := rest.ParseUint64OrReturnBadRequest(w, params.Get("val_id"))
-		if !ok {
-			return
-		}
-
-		slashAmount, ok := big.NewInt(0).SetString(params.Get("slash_amount"), 10)
-		if !ok {
-			return
-		}
-
-		RestLogger.Info("Slashing validator - ", "valID", valID, "slashAmount", slashAmount)
-
-		slashingParams := types.NewValidatorSlashParams(hmTypes.ValidatorID(valID), slashAmount)
-
-		bz, err := cliCtx.Codec.MarshalJSON(slashingParams)
-		if err != nil {
-			RestLogger.Info("Error marshallingSlashing Params - ", "err", err)
-			rest.WriteErrorResponse(w, http.StatusBadRequest, err.Error())
-			return
-		}
-
-		res, height, err := cliCtx.QueryWithData(fmt.Sprintf("custom/%s/%s", types.QuerierRoute, types.QuerySlashValidator), bz)
-		if err != nil {
-			RestLogger.Info("Error query data - ", "err", err)
-			rest.WriteErrorResponse(w, http.StatusBadRequest, err.Error())
-			return
-		}
-
-		RestLogger.Info("Slashed validator successfully ", "res", res)
-
-		cliCtx.WithHeight(height)
-		rest.PostProcessResponse(w, cliCtx, res)
-	}
-}
-
-// Returns Dividend Account information by ID
-func dividendAccountByIDHandlerFn(cliCtx context.CLIContext) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		vars := mux.Vars(r)
-
-		cliCtx, ok := rest.ParseQueryHeightOrReturnBadRequest(w, cliCtx, r)
-		if !ok {
-			return
-		}
-
-		// get id
-		id, ok := rest.ParseUint64OrReturnBadRequest(w, vars["id"])
-		if !ok {
-			return
-		}
-
-		// get query params
-		queryParams, err := cliCtx.Codec.MarshalJSON(types.NewQueryDividendAccountParams(hmTypes.DividendAccountID(id)))
-		if err != nil {
-			hmRest.WriteErrorResponse(w, http.StatusInternalServerError, err.Error())
-			return
-		}
-
-		res, height, err := cliCtx.QueryWithData(fmt.Sprintf("custom/%s/%s", types.QuerierRoute, types.QueryDividendAccount), queryParams)
-		if err != nil {
-			RestLogger.Error("Error while fetching Dividend account", "Error", err.Error())
-			hmRest.WriteErrorResponse(w, http.StatusBadRequest, err.Error())
-			return
-		}
-
-		// error if no dividend account found
-		if ok := hmRest.ReturnNotFoundIfNoContent(w, res, "No Dividend Account found"); !ok {
-			return
-		}
-
-		// return result
-		cliCtx = cliCtx.WithHeight(height)
-		rest.PostProcessResponse(w, cliCtx, res)
-	}
-}
-
-// dividendAccountRootHandlerFn returns genesis accountroothash
-func dividendAccountRootHandlerFn(
-	cliCtx context.CLIContext,
-) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		cliCtx, ok := rest.ParseQueryHeightOrReturnBadRequest(w, cliCtx, r)
-		if !ok {
-			return
-		}
-
-		res, _, err := cliCtx.QueryWithData(fmt.Sprintf("custom/%s/%s", types.QuerierRoute, types.QueryDividendAccountRoot), nil)
-		RestLogger.Debug("accountRootHash querier response", "res", res)
-
-		if err != nil {
-			RestLogger.Error("Error while calculating dividend AccountRoot  ", "Error", err.Error())
-			rest.WriteErrorResponse(w, http.StatusBadRequest, err.Error())
-			return
-		}
-
-		// error if no checkpoint found
-		if ok := hmRest.ReturnNotFoundIfNoContent(w, res, "Dividend AccountRoot not found"); !ok {
-			RestLogger.Error("AccountRoot not found ", "Error", err.Error())
-			return
-		}
-
-		var accountRootHash = hmTypes.BytesToHeimdallHash(res)
-		RestLogger.Debug("Fetched Dividend accountRootHash ", "AccountRootHash", accountRootHash)
-
-		result, err := json.Marshal(&accountRootHash)
-		if err != nil {
-			RestLogger.Error("Error while marshalling response to Json", "error", err)
-			rest.WriteErrorResponse(w, http.StatusBadRequest, err.Error())
-			return
-		}
-
-		// return result
-
-		rest.PostProcessResponse(w, cliCtx, result)
-	}
-}
-
 // Returns proposer Bonus Percent information
 func proposerBonusPercentHandlerFn(cliCtx context.CLIContext) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
@@ -452,99 +303,6 @@ func proposerBonusPercentHandlerFn(cliCtx context.CLIContext) http.HandlerFunc {
 			return
 		}
 		rest.PostProcessResponse(w, cliCtx, result)
-	}
-}
-
-// Returns Merkle path for dividendAccountID using dividend Account Tree
-func dividendAccountProofHandlerFn(cliCtx context.CLIContext) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-
-		vars := mux.Vars(r)
-
-		cliCtx, ok := rest.ParseQueryHeightOrReturnBadRequest(w, cliCtx, r)
-		if !ok {
-			return
-		}
-
-		// get id
-		id, ok := rest.ParseUint64OrReturnBadRequest(w, vars["id"])
-		if !ok {
-			return
-		}
-
-		// get query params
-		queryParams, err := cliCtx.Codec.MarshalJSON(types.NewQueryAccountProofParams(hmTypes.DividendAccountID(id)))
-		if err != nil {
-			hmRest.WriteErrorResponse(w, http.StatusInternalServerError, err.Error())
-			return
-		}
-
-		res, height, err := cliCtx.QueryWithData(fmt.Sprintf("custom/%s/%s", types.QuerierRoute, types.QueryAccountProof), queryParams)
-		if err != nil {
-			RestLogger.Error("Error while fetching merkle proof", "Error", err.Error())
-			hmRest.WriteErrorResponse(w, http.StatusBadRequest, err.Error())
-			return
-		}
-
-		// error if account proof  not found
-		if ok := hmRest.ReturnNotFoundIfNoContent(w, res, "No proof found"); !ok {
-			return
-		}
-
-		// return result
-		cliCtx = cliCtx.WithHeight(height)
-		rest.PostProcessResponse(w, cliCtx, res)
-	}
-}
-
-// VerifyAccountProofHandlerFn - Returns true if given Merkle path for dividendAccountID is valid
-func VerifyAccountProofHandlerFn(cliCtx context.CLIContext) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		cliCtx, ok := rest.ParseQueryHeightOrReturnBadRequest(w, cliCtx, r)
-		if !ok {
-			return
-		}
-
-		params := r.URL.Query()
-		id, ok := rest.ParseUint64OrReturnBadRequest(w, params.Get("id"))
-		if !ok {
-			return
-		}
-
-		accountProof := params.Get("proof")
-
-		RestLogger.Info("Verify Account Proof- ", "valID", id, "accountProof", accountProof)
-
-		// get query params
-		queryParams, err := cliCtx.Codec.MarshalJSON(types.NewQueryVerifyAccountProofParams(hmTypes.DividendAccountID(id), accountProof))
-		if err != nil {
-			hmRest.WriteErrorResponse(w, http.StatusInternalServerError, err.Error())
-			return
-		}
-
-		res, height, err := cliCtx.QueryWithData(fmt.Sprintf("custom/%s/%s", types.QuerierRoute, types.QueryVerifyAccountProof), queryParams)
-		if err != nil {
-			RestLogger.Error("Error while verifying merkle proof", "Error", err.Error())
-			hmRest.WriteErrorResponse(w, http.StatusBadRequest, err.Error())
-			return
-		}
-
-		var accountProofStatus bool
-		if err := json.Unmarshal(res, &accountProofStatus); err != nil {
-			hmRest.WriteErrorResponse(w, http.StatusBadRequest, err.Error())
-			return
-		}
-
-		res, err = json.Marshal(map[string]interface{}{"result": accountProofStatus})
-		if err != nil {
-			hmRest.WriteErrorResponse(w, http.StatusBadRequest, err.Error())
-			return
-		}
-
-		// return result
-		cliCtx = cliCtx.WithHeight(height)
-		rest.PostProcessResponse(w, cliCtx, res)
-
 	}
 }
