+++ staking/querier.go
@@ -1,7 +1,6 @@
 package staking
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
 	"math/big"
@@ -11,7 +10,6 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	abci "github.com/tendermint/tendermint/abci/types"
 
-	checkpointTypes "github.com/maticnetwork/heimdall/checkpoint/types"
 	"github.com/maticnetwork/heimdall/helper"
 	"github.com/maticnetwork/heimdall/staking/types"
 	hmTypes "github.com/maticnetwork/heimdall/types"
@@ -33,14 +31,6 @@ func NewQuerier(keeper Keeper, contractCaller helper.IContractCaller) sdk.Querie
 			return handleQueryProposer(ctx, req, keeper)
 		case types.QueryCurrentProposer:
 			return handleQueryCurrentProposer(ctx, req, keeper)
-		case types.QueryDividendAccount:
-			return handleQueryDividendAccount(ctx, req, keeper)
-		case types.QueryDividendAccountRoot:
-			return handleDividendAccountRoot(ctx, req, keeper)
-		case types.QueryAccountProof:
-			return handleQueryAccountProof(ctx, req, keeper, contractCaller)
-		case types.QueryVerifyAccountProof:
-			return handleQueryVerifyAccountProof(ctx, req, keeper)
 		case types.QueryStakingSequence:
 			return handleQueryStakingSequence(ctx, req, keeper, contractCaller)
 
@@ -153,98 +143,6 @@ func handleQueryProposer(ctx sdk.Context, req abci.RequestQuery, keeper Keeper) 
 func handleQueryCurrentProposer(ctx sdk.Context, req abci.RequestQuery, keeper Keeper) ([]byte, sdk.Error) {
 	proposer := keeper.GetCurrentProposer(ctx)
 	bz, err := json.Marshal(proposer)
-	if err != nil {
-		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not marshal result to JSON", err.Error()))
-	}
-	return bz, nil
-}
-
-func handleQueryDividendAccount(ctx sdk.Context, req abci.RequestQuery, keeper Keeper) ([]byte, sdk.Error) {
-	var params types.QueryDividendAccountParams
-	if err := keeper.cdc.UnmarshalJSON(req.Data, &params); err != nil {
-		return nil, sdk.ErrInternal(fmt.Sprintf("failed to parse params: %s", err))
-	}
-
-	// get dividend account info
-	dividendAccount, err := keeper.GetDividendAccountByID(ctx, params.DividendAccountID)
-	if err != nil {
-		return nil, sdk.ErrUnknownRequest("No dividend account found")
-	}
-
-	// json record
-	bz, err := json.Marshal(dividendAccount)
-	if err != nil {
-		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not marshal result to JSON", err.Error()))
-	}
-	return bz, nil
-}
-
-func handleDividendAccountRoot(ctx sdk.Context, req abci.RequestQuery, keeper Keeper) ([]byte, sdk.Error) {
-	// Calculate new account root hash
-	dividendAccounts := keeper.GetAllDividendAccounts(ctx)
-	accountRoot, err := checkpointTypes.GetAccountRootHash(dividendAccounts)
-	if err != nil {
-		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not fetch accountroothash ", err.Error()))
-	}
-	return accountRoot, nil
-}
-
-func handleQueryAccountProof(ctx sdk.Context, req abci.RequestQuery, keeper Keeper, contractCallerObj helper.IContractCaller) ([]byte, sdk.Error) {
-	// 1. Fetch AccountRoot a1 present on RootChainContract
-	// 2. Fetch AccountRoot a2 from current account
-	// 3. if a1 == a2, Calculate merkle path using GetAllDividendAccounts
-
-	var params types.QueryAccountProofParams
-	if err := keeper.cdc.UnmarshalJSON(req.Data, &params); err != nil {
-		return nil, sdk.ErrInternal(fmt.Sprintf("failed to parse params: %s", err))
-	}
-
-	chainParams := keeper.chainKeeper.GetParams(ctx)
-
-	stakingInfoAddress := chainParams.ChainParams.StakingInfoAddress.EthAddress()
-	stakingInfoInstance, _ := contractCallerObj.GetStakingInfoInstance(stakingInfoAddress)
-
-	accountRootOnChain, err := contractCallerObj.CurrentAccountStateRoot(stakingInfoInstance)
-	if err != nil {
-		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not fetch account root from onchain ", err.Error()))
-	}
-
-	dividendAccounts := keeper.GetAllDividendAccounts(ctx)
-	currentStateAccountRoot, err := checkpointTypes.GetAccountRootHash(dividendAccounts)
-
-	if bytes.Equal(accountRootOnChain[:], currentStateAccountRoot) {
-		// Calculate new account root hash
-		merkleProof, index, _ := checkpointTypes.GetAccountProof(dividendAccounts, params.DividendAccountID)
-		accountProof := hmTypes.NewDividendAccountProof(params.DividendAccountID, merkleProof, index)
-		// json record
-		bz, err := json.Marshal(accountProof)
-		if err != nil {
-			return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not marshal result to JSON", err.Error()))
-		}
-		return bz, nil
-
-	} else {
-		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not fetch merkle proof ", err.Error()))
-	}
-}
-
-func handleQueryVerifyAccountProof(ctx sdk.Context, req abci.RequestQuery, keeper Keeper) ([]byte, sdk.Error) {
-
-	var params types.QueryVerifyAccountProofParams
-	if err := keeper.cdc.UnmarshalJSON(req.Data, &params); err != nil {
-		return nil, sdk.ErrInternal(fmt.Sprintf("failed to parse params: %s", err))
-	}
-
-	dividendAccounts := keeper.GetAllDividendAccounts(ctx)
-
-	// Verify account proof
-	accountProofStatus, err := checkpointTypes.VerifyAccountProof(dividendAccounts, params.DividendAccountID, params.AccountProof)
-	if err != nil {
-		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not verify merkle proof ", err.Error()))
-	}
-
-	// json record
-	bz, err := json.Marshal(accountProofStatus)
 	if err != nil {
 		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not marshal result to JSON", err.Error()))
 	}
