+++ staking/types/querier.go
@@ -16,11 +16,6 @@ const (
 	QueryProposer             = "proposer"
 	QueryCurrentProposer      = "current-proposer"
 	QueryProposerBonusPercent = "proposer-bonus-percent"
-	QueryDividendAccount      = "dividend-account"
-	QueryDividendAccountRoot  = "dividend-account-root"
-	QueryAccountProof         = "dividend-account-proof"
-	QueryVerifyAccountProof   = "verify-account-proof"
-	QuerySlashValidator       = "slash-validator"
 	QueryStakingSequence      = "staking-sequence"
 )
 
@@ -42,37 +37,6 @@ type QueryValidatorParams struct {
 // NewQueryValidatorParams creates a new instance of QueryValidatorParams.
 func NewQueryValidatorParams(validatorID types.ValidatorID) QueryValidatorParams {
 	return QueryValidatorParams{ValidatorID: validatorID}
-}
-
-// QueryDividendAccountParams defines the params for querying dividend account status.
-type QueryDividendAccountParams struct {
-	DividendAccountID types.DividendAccountID `json:"dividend_account_id"`
-}
-
-// NewQueryDividendAccountParams creates a new instance of QueryDividendAccountParams.
-func NewQueryDividendAccountParams(dividendAccountID types.DividendAccountID) QueryDividendAccountParams {
-	return QueryDividendAccountParams{DividendAccountID: dividendAccountID}
-}
-
-// QueryAccountProofParams defines the params for querying account proof.
-type QueryAccountProofParams struct {
-	DividendAccountID types.DividendAccountID `json:"dividend_account_id"`
-}
-
-// NewQueryAccountProofParams creates a new instance of QueryAccountProofParams.
-func NewQueryAccountProofParams(dividendAccountID types.DividendAccountID) QueryAccountProofParams {
-	return QueryAccountProofParams{DividendAccountID: dividendAccountID}
-}
-
-// QueryVerifyAccountProofParams defines the params for verifying account proof.
-type QueryVerifyAccountProofParams struct {
-	DividendAccountID types.DividendAccountID `json:"dividend_account_id"`
-	AccountProof      string                  `json:"account_proof"`
-}
-
-// NewQueryVerifyAccountProofParams creates a new instance of QueryVerifyAccountProofParams.
-func NewQueryVerifyAccountProofParams(dividendAccountID types.DividendAccountID, accountProof string) QueryVerifyAccountProofParams {
-	return QueryVerifyAccountProofParams{DividendAccountID: dividendAccountID, AccountProof: accountProof}
 }
 
 // QueryProposerParams defines the params for querying val status.
