+++ topup/keeper.go
@@ -1,6 +1,10 @@
 package topup
 
 import (
+	"encoding/hex"
+	"errors"
+	"math/big"
+
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
@@ -9,6 +13,7 @@ import (
 	"github.com/maticnetwork/heimdall/params/subspace"
 	"github.com/maticnetwork/heimdall/staking"
 	"github.com/maticnetwork/heimdall/topup/types"
+	hmTypes "github.com/maticnetwork/heimdall/types"
 	"github.com/tendermint/tendermint/libs/log"
 )
 
@@ -17,6 +22,8 @@ var (
 	DefaultValue = []byte{0x01}
 	// TopupSequencePrefixKey represents topup sequence prefix key
 	TopupSequencePrefixKey = []byte{0x81}
+
+	DividendAccountMapKey = []byte{0x82} // prefix for each key for Dividend Account Map
 )
 
 // Keeper stores all related data
@@ -115,4 +122,109 @@ func (keeper Keeper) SetTopupSequence(ctx sdk.Context, sequence string) {
 func (keeper Keeper) HasTopupSequence(ctx sdk.Context, sequence string) bool {
 	store := ctx.KVStore(keeper.key)
 	return store.Has(GetTopupSequenceKey(sequence))
+}
+
+// GetDividendAccountMapKey returns dividend account map
+func GetDividendAccountMapKey(id []byte) []byte {
+	return append(DividendAccountMapKey, id...)
+}
+
+// AddDividendAccount adds DividendAccount index with DividendID
+func (k *Keeper) AddDividendAccount(ctx sdk.Context, dividendAccount hmTypes.DividendAccount) error {
+	store := ctx.KVStore(k.key)
+	// marshall dividend account
+	bz, err := hmTypes.MarshallDividendAccount(k.cdc, dividendAccount)
+	if err != nil {
+		return err
+	}
+
+	store.Set(GetDividendAccountMapKey(dividendAccount.ID.Bytes()), bz)
+	k.Logger(ctx).Debug("DividendAccount Stored", "key", hex.EncodeToString(GetDividendAccountMapKey(dividendAccount.ID.Bytes())), "dividendAccount", dividendAccount.String())
+	return nil
+}
+
+// GetDividendAccountByID will return DividendAccount of valID
+func (k *Keeper) GetDividendAccountByID(ctx sdk.Context, dividendID hmTypes.DividendAccountID) (dividendAccount hmTypes.DividendAccount, err error) {
+
+	// check if dividend account exists
+	if !k.CheckIfDividendAccountExists(ctx, dividendID) {
+		return dividendAccount, errors.New("Dividend Account not found")
+	}
+
+	// Get DividendAccount key
+	store := ctx.KVStore(k.key)
+	key := GetDividendAccountMapKey(dividendID.Bytes())
+
+	// unmarshall dividend account and return
+	dividendAccount, err = hmTypes.UnMarshallDividendAccount(k.cdc, store.Get(key))
+	if err != nil {
+		return dividendAccount, err
+	}
+
+	return dividendAccount, nil
+}
+
+// CheckIfDividendAccountExists will return true if dividend account exists
+func (k *Keeper) CheckIfDividendAccountExists(ctx sdk.Context, dividendID hmTypes.DividendAccountID) (ok bool) {
+	store := ctx.KVStore(k.key)
+	key := GetDividendAccountMapKey(dividendID.Bytes())
+	return store.Has(key)
+}
+
+// GetAllDividendAccounts returns all DividendAccountss
+func (k *Keeper) GetAllDividendAccounts(ctx sdk.Context) (dividendAccounts []hmTypes.DividendAccount) {
+	// iterate through dividendAccounts and create dividendAccounts update array
+	k.IterateDividendAccountsByPrefixAndApplyFn(ctx, DividendAccountMapKey, func(dividendAccount hmTypes.DividendAccount) error {
+		// append to list of dividendUpdates
+		dividendAccounts = append(dividendAccounts, dividendAccount)
+		return nil
+	})
+
+	return
+}
+
+// AddFeeToDividendAccount adds fee to dividend account for withdrawal
+func (k *Keeper) AddFeeToDividendAccount(ctx sdk.Context, valID hmTypes.ValidatorID, fee *big.Int) sdk.Error {
+	// Get or create dividend account
+	var dividendAccount hmTypes.DividendAccount
+
+	if k.CheckIfDividendAccountExists(ctx, hmTypes.DividendAccountID(valID)) {
+		dividendAccount, _ = k.GetDividendAccountByID(ctx, hmTypes.DividendAccountID(valID))
+	} else {
+		dividendAccount = hmTypes.DividendAccount{
+			ID:            hmTypes.DividendAccountID(valID),
+			FeeAmount:     big.NewInt(0).String(),
+			SlashedAmount: big.NewInt(0).String(),
+		}
+	}
+
+	// update fee
+	oldFee, _ := big.NewInt(0).SetString(dividendAccount.FeeAmount, 10)
+	totalFee := big.NewInt(0).Add(oldFee, fee).String()
+	dividendAccount.FeeAmount = totalFee
+
+	k.Logger(ctx).Info("Dividend Account fee of validator ", "ID", dividendAccount.ID, "Fee", dividendAccount.FeeAmount)
+	if err := k.AddDividendAccount(ctx, dividendAccount); err != nil {
+		k.Logger(ctx).Error("AddFeeToDividendAccount | AddDividendAccount", "error", err)
+	}
+	return nil
+}
+
+// IterateDividendAccountsByPrefixAndApplyFn iterate dividendAccounts and apply the given function.
+func (k *Keeper) IterateDividendAccountsByPrefixAndApplyFn(ctx sdk.Context, prefix []byte, f func(dividendAccount hmTypes.DividendAccount) error) {
+	store := ctx.KVStore(k.key)
+
+	// get validator iterator
+	iterator := sdk.KVStorePrefixIterator(store, prefix)
+	defer iterator.Close()
+
+	// loop through dividendAccounts
+	for ; iterator.Valid(); iterator.Next() {
+		// unmarshall dividendAccount
+		dividendAccount, _ := hmTypes.UnMarshallDividendAccount(k.cdc, iterator.Value())
+		// call function and return if required
+		if err := f(dividendAccount); err != nil {
+			return
+		}
+	}
 }
