+++ topup/querier.go
@@ -1,12 +1,15 @@
 package topup
 
 import (
+	"bytes"
+	"encoding/json"
 	"fmt"
 	"math/big"
 	"time"
 
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	checkpointTypes "github.com/maticnetwork/heimdall/checkpoint/types"
 	"github.com/maticnetwork/heimdall/helper"
 	"github.com/maticnetwork/heimdall/topup/types"
 	hmTypes "github.com/maticnetwork/heimdall/types"
@@ -19,6 +22,14 @@ func NewQuerier(k Keeper, contractCaller helper.IContractCaller) sdk.Querier {
 		switch path[0] {
 		case types.QuerySequence:
 			return querySequence(ctx, req, k, contractCaller)
+		case types.QueryDividendAccount:
+			return handleQueryDividendAccount(ctx, req, k)
+		case types.QueryDividendAccountRoot:
+			return handleDividendAccountRoot(ctx, req, k)
+		case types.QueryAccountProof:
+			return handleQueryAccountProof(ctx, req, k, contractCaller)
+		case types.QueryVerifyAccountProof:
+			return handleQueryVerifyAccountProof(ctx, req, k)
 		default:
 			return nil, sdk.ErrUnknownRequest("unknown topup query endpoint")
 		}
@@ -56,5 +67,97 @@ func querySequence(ctx sdk.Context, req abci.RequestQuery, k Keeper, contractCal
 		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not marshal result to JSON", err.Error()))
 	}
 
+	return bz, nil
+}
+
+func handleQueryDividendAccount(ctx sdk.Context, req abci.RequestQuery, keeper Keeper) ([]byte, sdk.Error) {
+	var params types.QueryDividendAccountParams
+	if err := keeper.cdc.UnmarshalJSON(req.Data, &params); err != nil {
+		return nil, sdk.ErrInternal(fmt.Sprintf("failed to parse params: %s", err))
+	}
+
+	// get dividend account info
+	dividendAccount, err := keeper.GetDividendAccountByID(ctx, params.DividendAccountID)
+	if err != nil {
+		return nil, sdk.ErrUnknownRequest("No dividend account found")
+	}
+
+	// json record
+	bz, err := json.Marshal(dividendAccount)
+	if err != nil {
+		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not marshal result to JSON", err.Error()))
+	}
+	return bz, nil
+}
+
+func handleDividendAccountRoot(ctx sdk.Context, req abci.RequestQuery, keeper Keeper) ([]byte, sdk.Error) {
+	// Calculate new account root hash
+	dividendAccounts := keeper.GetAllDividendAccounts(ctx)
+	accountRoot, err := checkpointTypes.GetAccountRootHash(dividendAccounts)
+	if err != nil {
+		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not fetch accountroothash ", err.Error()))
+	}
+	return accountRoot, nil
+}
+
+func handleQueryAccountProof(ctx sdk.Context, req abci.RequestQuery, keeper Keeper, contractCallerObj helper.IContractCaller) ([]byte, sdk.Error) {
+	// 1. Fetch AccountRoot a1 present on RootChainContract
+	// 2. Fetch AccountRoot a2 from current account
+	// 3. if a1 == a2, Calculate merkle path using GetAllDividendAccounts
+
+	var params types.QueryAccountProofParams
+	if err := keeper.cdc.UnmarshalJSON(req.Data, &params); err != nil {
+		return nil, sdk.ErrInternal(fmt.Sprintf("failed to parse params: %s", err))
+	}
+
+	chainParams := keeper.chainKeeper.GetParams(ctx)
+
+	stakingInfoAddress := chainParams.ChainParams.StakingInfoAddress.EthAddress()
+	stakingInfoInstance, _ := contractCallerObj.GetStakingInfoInstance(stakingInfoAddress)
+
+	accountRootOnChain, err := contractCallerObj.CurrentAccountStateRoot(stakingInfoInstance)
+	if err != nil {
+		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not fetch account root from onchain ", err.Error()))
+	}
+
+	dividendAccounts := keeper.GetAllDividendAccounts(ctx)
+	currentStateAccountRoot, err := checkpointTypes.GetAccountRootHash(dividendAccounts)
+
+	if bytes.Equal(accountRootOnChain[:], currentStateAccountRoot) {
+		// Calculate new account root hash
+		merkleProof, index, _ := checkpointTypes.GetAccountProof(dividendAccounts, params.DividendAccountID)
+		accountProof := hmTypes.NewDividendAccountProof(params.DividendAccountID, merkleProof, index)
+		// json record
+		bz, err := json.Marshal(accountProof)
+		if err != nil {
+			return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not marshal result to JSON", err.Error()))
+		}
+		return bz, nil
+
+	} else {
+		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not fetch merkle proof ", err.Error()))
+	}
+}
+
+func handleQueryVerifyAccountProof(ctx sdk.Context, req abci.RequestQuery, keeper Keeper) ([]byte, sdk.Error) {
+
+	var params types.QueryVerifyAccountProofParams
+	if err := keeper.cdc.UnmarshalJSON(req.Data, &params); err != nil {
+		return nil, sdk.ErrInternal(fmt.Sprintf("failed to parse params: %s", err))
+	}
+
+	dividendAccounts := keeper.GetAllDividendAccounts(ctx)
+
+	// Verify account proof
+	accountProofStatus, err := checkpointTypes.VerifyAccountProof(dividendAccounts, params.DividendAccountID, params.AccountProof)
+	if err != nil {
+		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not verify merkle proof ", err.Error()))
+	}
+
+	// json record
+	bz, err := json.Marshal(accountProofStatus)
+	if err != nil {
+		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not marshal result to JSON", err.Error()))
+	}
 	return bz, nil
 }
