+++ topup/simulation/genesis.go
@@ -2,11 +2,13 @@ package simulation
 
 import (
 	"fmt"
+	"math/big"
 	"math/rand"
 	"strconv"
 
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/maticnetwork/heimdall/topup/types"
+	hmTypes "github.com/maticnetwork/heimdall/types"
 	"github.com/maticnetwork/heimdall/types/module"
 	"github.com/maticnetwork/heimdall/types/simulation"
 )
@@ -22,6 +24,7 @@ func GenSequenceNumber(r *rand.Rand) string {
 // RandomizeGenState returns topup genesis
 func RandomizeGenState(simState *module.SimulationState) {
 	var sequences []string
+	dividendAccounts := make([]hmTypes.DividendAccount, 5)
 
 	for i := 0; i < 5; i++ {
 		var sequenceNumber string
@@ -31,9 +34,16 @@ func RandomizeGenState(simState *module.SimulationState) {
 			},
 		)
 		sequences = append(sequences, sequenceNumber)
+
+		// create dividend account for validator
+		dividendAccounts[i] = hmTypes.NewDividendAccount(
+			hmTypes.NewDividendAccountID(uint64(i)),
+			big.NewInt(0).String(),
+			big.NewInt(0).String(),
+		)
 	}
 
-	topupGenesis := types.NewGenesisState(sequences)
+	topupGenesis := types.NewGenesisState(sequences, dividendAccounts)
 	fmt.Printf("Selected randomly generated topup sequences:\n%s\n", codec.MustMarshalJSONIndent(simState.Cdc, topupGenesis))
 	simState.GenState[types.ModuleName] = simState.Cdc.MustMarshalJSON(topupGenesis)
 }
